factor out task computation code from task execution
